requirements.txt
blosc2==2.6.2
greenlet==3.0.3
msgpack==1.0.8
ndindex==1.8
numexpr==2.10.0
numpy==1.26.4
overpy==0.7
packaging==24.1
pandas==2.2.2
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
simpledbf==0.2.6
six==1.16.0
SQLAlchemy==2.0.30
tables==3.9.2
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
